# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ThreadCap API
  version: 1.0.0
  description: >-
    ThreadCap REST API. Product spec: docs/04-api-contracts.md.
    Errors use the ThreadCap envelope {requestId, error:{code,message,details}}.
    Auth: Authorization: Bearer <jwt> (web/extension);
    X-API-Key: cht_... (MCP/skills). Conventions per docs/00-index.md.
servers:
  - url: https://api.threadcap.app/v1
  - url: http://localhost:3000/v1
...
paths:
...
  /capsules:
    get:
      tags: [Capsules]
      operationId: listCapsules
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: team, in: query, schema: { type: string } }
        - { name: project, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, archived, deleted], default: active } }
        - { name: mode, in: query, schema: { type: string, enum: [raw, smart] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paged capsule list
          content:
            application/json:
              schema:
                type: object
                required: [items, meta]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Capsule' }
                  meta: { $ref: '#/components/schemas/Meta' }
    post:
      tags: [Capsules]
      operationId: createCapsule
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content, source]
              properties:
                name: { type: string, maxLength: 120 }
                description: { type: string, maxLength: 500 }
                teamId: { type: string, nullable: true }
                projectId: { type: string, nullable: true }
                tags:
                  type: array
                  maxItems: 20
                  items: { type: string }
                source: { $ref: '#/components/schemas/Source' }
                content: { $ref: '#/components/schemas/CapsuleContent' }
      responses:
        '201':
          description: Created capsule + v1
          content:
            application/json:
              schema:
                type: object
                required: [capsule, version]
                properties:
                  capsule: { $ref: '#/components/schemas/Capsule' }
                  version: { $ref: '#/components/schemas/CapsuleVersion' }
        '409': { $ref: '#/components/responses/Error' }
  /capsules/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, example: cap_01HZX... } }
    get:
      tags: [Capsules]
      operationId: getCapsule
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Capsule wrapper + latestVersion content + live sections
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capsule' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      tags: [Capsules]
      operationId: updateCapsule
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                description: { type: string, maxLength: 500 }
                teamId: { type: string, nullable: true }
                projectId: { type: string, nullable: true }
                tags: { type: array, maxItems: 20, items: { type: string } }
                status: { type: string, enum: [active, archived] }
      responses:
        '200': { description: Updated capsule, body Capsule }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Capsules]
      operationId: deleteCapsule
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Soft-deleted (purged after 30 days) }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    Error:
      description: ThreadCap error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [requestId, error]
      properties:
        requestId: { type: string, example: req_01HZX... }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: CAPSULE_NOT_FOUND }
            message: { type: string }
            details:
              oneOf:
                - { type: object }
                - { type: 'null' }
    Meta:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        total: { type: integer }
...
    Source:
      type: object
      required: [platform]
      properties:
        platform: { type: string, enum: [chatgpt, claude, gemini, gmail, terminal, web, manual] }
        projectRef: { type: string, nullable: true }
        capturedAt: { type: string, format: date-time }
    Message:
      type: object
      required: [role, content, ts]
      properties:
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string }
        ts: { type: string, format: date-time }
        sourceTool: { type: string }
        tokenEstimate: { type: integer }
    ContextItem:
      type: object
      required: [id, text]
      properties:
        id: { type: string }
        text: { type: string }
        status: { type: string, enum: [open, accepted, superseded, rejected] }
    CapsuleContent:
      type: object
      required: [objective, background, requirements, constraints, decisions, assumptions, openQuestions, conversation, attachments, links, summary]
      properties:
        objective: { type: string }
        background: { type: string }
        requirements: { type: array, items: { $ref: '#/components/schemas/ContextItem' } }
        constraints: { type: array, items: { $ref: '#/components/schemas/ContextItem' } }
        decisions: { type: array, items: { $ref: '#/components/schemas/ContextItem' } }
        assumptions: { type: array, items: { $ref: '#/components/schemas/ContextItem' } }
        openQuestions: { type: array, items: { $ref: '#/components/schemas/ContextItem' } }
        conversation: { type: array, items: { $ref: '#/components/schemas/Message' } }
        attachments:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              mimeType: { type: string }
              size: { type: integer }
              role: { type: string, enum: [source, generated, manual] }
        links:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              label: { type: string }
        summary: { type: string }
    CapsuleVersion:
      type: object
      required: [id, capsuleId, versionNumber, changeSummary, content, contentHash, createdAt]
      properties:
        id: { type: string }
        capsuleId: { type: string }
        versionNumber: { type: integer }
        parentVersionId: { type: string, nullable: true }
        branchId: { type: string, nullable: true }
        authorId: { type: string }
        changeSummary: { type: string }
        content: { $ref: '#/components/schemas/CapsuleContent' }
        contentHash: { type: string }
        tokenEstimate: { type: integer }
        isCurrent: { type: boolean }
        createdAt: { type: string, format: date-time }
    Capsule:
      type: object
      required: [id, workspaceId, name, status, source, versionCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        teamId: { type: string, nullable: true }
        projectId: { type: string, nullable: true }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived, deleted] }
        source: { $ref: '#/components/schemas/Source' }
        tags: { type: array, items: { type: string } }
        primaryTag: { type: string, nullable: true }
        versionCount: { type: integer }
        currentVersionId: { type: string, nullable: true }
        latestVersion: { $ref: '#/components/schemas/CapsuleVersion' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...